requirements.txt
google-generativeai>=0.7.1
streamlit>=1.36.0
opencv-python>=4.10.0.84
mediapipe>=0.10